[Refactor] Remove IndexType template param from EigenTensor and unify to int64_t
Remove the `typename IndexType = Eigen::DenseIndex` template parameter from
EigenTensor, EigenMatrix, EigenVector, EigenScalar structs in both
paddle/phi and paddle/fluid namespaces. All Eigen::TensorMap and
Eigen::Tensor types now consistently use int64_t as index type.
Also update Tensor32BitIndex/To32BitDims to use int64_t.

+++ paddle/phi/kernels/impl/lu_kernel_impl.h
@@ -319,7 +319,7 @@ void SliceCompute(const Context& dev_ctx,
   if (in->numel() <= Eigen::NumTraits<int>::highest()) {
     // similar to tf.slice:
     // if element number less than INT_MAX, change the type of index to int
-    Eigen::DSizes<int, D> offsets_32bit, extents_32bit;
+    Eigen::DSizes<int64_t, D> offsets_32bit, extents_32bit;
     for (size_t i = 0; i < D; i++) {
       offsets_32bit[i] = offsets[i];
       extents_32bit[i] = extents[i];
